quals/misc/ilynaga/source/requirements.txt
deepface
scikit-image
transformers
gradio
pillow
tf-keras
torch